Add Data Mechanics Delight to Spark third-party projects
Data Mechanics Delight is a free, hosted, cross-platform monitoring tool for Spark. It's powered by an open-source Spark agent which works in any type of Spark infrastructure setup. 

The project is live as of November 2020 and will get significant improvement over the following weeks -- right now it is a dashboard giving you access to the Spark UI (Spark History Server), but we will soon add new debugging screens to complement and improve on the Spark UI.

We hope the Spark community will really benefit from this tool, and so we'd love to be listed on this page.

Reference: 
- https://www.kdnuggets.com/2020/07/monitoring-apache-spark-better-ui.html
- https://www.datamechanics.co/delight